fix(clean methods): move clean methods outside ViewBuilder class
